requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
breadability==0.1.20
celery==5.6.2
certifi==2026.2.25
chardet==7.2.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.3
djangorestframework==3.16.1
docopt==0.6.2
docopt-ng==0.9.0
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
lxml==6.0.2
lxml_html_clean==0.4.4
nltk==3.9.3
numpy==2.4.3
packaging==26.0
prompt_toolkit==3.0.52
pycountry==26.2.16
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rpds-py==0.30.0
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
sumy==0.12.0
textblob==0.19.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
